# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Consulting Revenue Platform API
  version: "1.0"
  description: |
    Multi-tenant API for time tracking, invoicing, Stripe payments, ledger, and reporting.
    Authenticated routes require `Authorization: Bearer <access_token>` and `X-Organization-ID: <uuid>`.
servers:
  - url: http://localhost:8080
    description: Local development
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access JWT from POST /auth/login or /auth/register.
...
paths:
...
  /v1/admin/reconciliation-summary:
    get:
      tags: [Admin]
      summary: Owner-only reconciliation of ledger payment captures vs paid invoice payments
      description: |
        Compares sums of `ledger_entries` with `event_type` payment_captured to sums of `payments.amount_minor`
        for invoices in `paid` status, per currency. Non-zero `drift_minor` indicates investigation is needed.
      security: [{ bearerAuth: [] }]
      parameters:
        - in: header
          name: X-Organization-ID
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Summary per currency
          content:
            application/json:
              schema:
                type: object
                properties:
                  currencies:
                    type: array
                    items:
                      type: object
                      properties:
                        currency:
                          type: string
                        ledger_payment_captured_minor:
                          type: integer
                          format: int64
                        paid_invoice_payments_minor:
                          type: integer
                          format: int64
                        drift_minor:
                          type: integer
                          format: int64
                        aligned:
                          type: boolean
        "403":
          description: Forbidden
...